requirements.txt
torch>=2.0.1
pillow>=10.0.0
numpy>=1.24.3
opencv-python>=4.7.0.72
tqdm>=4.65.0